requirements.txt
Django==6.0.1
gunicorn==25.1.0
whitenoise==6.12.0
dj-database-url==3.1.2
psycopg2-binary==2.9.11
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
python-dotenv==1.2.2
pytesseract==0.3.13
pillow==12.1.1
xhtml2pdf==0.2.17
pypdf==6.9.0
requests==2.32.5
RapidFuzz==3.14.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.3
PyYAML==6.0.3
isort==8.0.1
tzdata==2025.3
tzlocal==5.3.1
